refactor: study-room style 수정
