+++ src/components/studyroom/button/CreateButton.style.ts
@@ -2,5 +2,6 @@ import styled from "styled-components";
 
 export const CreateButtonStyle = styled.div`
   display: flex;
-  margin-left: 46px;
+  margin-left: auto;
+  margin-right: 29px;
 `;
